Rename payload sources to source
Sources and payloads should be named source and payload.
