+++ pyanaconda/modules/payloads/Makefile.am
@@ -14,7 +14,7 @@
 # You should have received a copy of the GNU Lesser General Public License
 # along with this program.  If not, see <http://www.gnu.org/licenses/>.
 
-SUBDIRS = base payload sources
+SUBDIRS = base payload source
 
 pkgpyexecdir = $(pyexecdir)/py$(PACKAGE_NAME)
 payload_moduledir = $(pkgpyexecdir)/modules/payloads
