+++ pyanaconda/modules/payloads/source/live_os/initialization.py
