+++ pyanaconda/modules/payloads/source/live_os/live_os_interface.py
@@ -21,7 +21,7 @@ from dasbus.server.interface import dbus_interface
 from dasbus.typing import *  # pylint: disable=wildcard-import
 from dasbus.server.property import emits_properties_changed
 from pyanaconda.modules.common.constants.interfaces import PAYLOAD_SOURCE_LIVE_OS
-from pyanaconda.modules.payloads.sources.source_base_interface import PayloadSourceBaseInterface
+from pyanaconda.modules.payloads.source.source_base_interface import PayloadSourceBaseInterface
 
 
 @dbus_interface(PAYLOAD_SOURCE_LIVE_OS.interface_name)
