+++ pyanaconda/modules/payloads/source/source_base.py
