+++ pyanaconda/modules/payloads/source/source_base_interface.py
